backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.46
fakeredis==2.20.0
redis==5.0.1
prometheus-fastapi-instrumentator==6.1.0
python-multipart==0.0.6
pydantic-settings==2.1.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator>=2.0.0
